requirements.txt
-f https://download.pytorch.org/whl/cu113/torch_stable.html
gdown==4.2.0
matplotlib==3.3.4
numpy==1.21.1
pandas==1.4.3
opencv_python==4.5.3.56
Pillow==8.4.0
torch==1.10.1+cu113
torchvision==0.11.2+cu113
tqdm==4.62.1
